[6058907][ONNX] Extract clear_types_and_shapes_recursive to autocast utils
Move the recursive type/shape clearing out of PrecisionConverter into
modelopt/onnx/autocast/utils.py as a shared clear_types_and_shapes_recursive()
next to walk_subgraphs_recursive; PrecisionConverter now delegates to it.
Behavior is unchanged (clear_shapes = not use_standalone_type_inference).

The standalone type-inference fallback on shape-inference failure already
lives in infer_types(), which the AutoCast path exercises via strict_mode=True.
